COM-494 Fix form logic not working for yes/no answers
The full name for answers "Yes" or "No" is "Yes full name" and "No full name". When applying the logic on a newly created form, the concept is identified by the full name, but when editing, the short name is needed.

This fix checks that the answer is either yes/no or yes full name/no full name so that both creation and edit of forms works.