COM-93 Refactor dispensation queries
Simplifies the query to calculate the ARV dispensed by category: dynamically retrieves the list of drugs from the ARV concept set and then calculates the total ARV dispensed for each category; all is done in one query. This approach reduces the number of reports sent to Excel from 7 to 1 which improves the rendering speed in Excel.

Simplifies the query prescribed_vs_dispensed : dynamically retrieves the list of drugs from the ARV concept set and then calculates the total prescribed and dispensed for each drug. This change makes the query simpler and cleaner by reducing the number of columns