COM-795 Fix condition
The function getObsDatetimeValueInSection returns NULL when the date doesn't exist; and the result of NULL BETWEEN value1 AND value is actually NULL (i.e. not 0 as I also intuitively expected it). The condition in the next line therefore also returns NULL, causing RETURN 0 not to be called.

Fixed it by checking if it's null instead of false.